+++ src/base/browser/basic/tree/multiTreeModel.ts
@@ -1,7 +1,7 @@
 import { Register } from "src/base/common/event";
 import { ISpliceable } from "src/base/common/range";
-import { IIndexTreeModelCreationOptions, IIndexTreeModel, IndexTreeModel, ITreeModelSpliceEvent, ITreeModelSpliceOptions } from "src/base/common/tree/indexTreeModel";
-import { ITreeModel, ITreeNode, ITreeNodeItem } from "src/base/common/tree/tree";
+import { IIndexTreeModelCreationOptions, IIndexTreeModel, IndexTreeModel, ITreeModelSpliceEvent, ITreeModelSpliceOptions } from "src/base/browser/basic/tree/indexTreeModel";
+import { ITreeModel, ITreeNode, ITreeNodeItem } from "src/base/browser/basic/tree/tree";
 
 /**
  * An interface only for {@link MultiTreeModel}.
