+++ src/base/browser/basic/tree/treeListWidget.ts
@@ -1,7 +1,7 @@
+import { ITreeNode } from "src/base/browser/basic/tree/tree";
 import { IListItemProvider } from "src/base/browser/secondary/listView/listItemProvider";
 import { IListViewRenderer } from "src/base/browser/secondary/listView/listRenderer";
 import { IListWidget, IListWidgetOpts, ListWidget } from "src/base/browser/secondary/listWidget/listWidget";
-import { ITreeNode } from "src/base/common/tree/tree";
 
 /**
  * An interface for {@link TreeListWidget}.
